test(react): give the JSDOM harness one home
Twelve UI test files each carried their own copy of the same harness: snapshot
globals, install them off a JSDOM window, stub ResizeObserver and matchMedia,
stub requestAnimationFrame, restore on teardown. Identical code, twelve times.

That is how one defect ended up in all twelve. The rAF stub is a real timer, and
nothing drained the frames still queued at teardown, so a frame React scheduled
just before unmount became a pending timer -- which the op sanitizer reports as
a leak, at suite level, with no step and no stack, on whatever branch happens to
be pushing.

Move the mechanics into one module and have each file call it. The helper owns
snapshot, install, drain and restore. It deliberately does not decide which
globals a test installs: callers opt in to matchMedia and to any extra window
constructors they need, so every file keeps exactly the globals it had and this
change alters no behaviour.

Net 798 lines removed, and no file stubs requestAnimationFrame by hand any more.

The export is named installComponentDom rather than installDomGlobals because
fourteen other test files already define a local function by that name with a
different signature. An import would have shadowed those declarations, and with
--no-check on the test tasks the suite would still have run green.
